requirements.txt
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
matplotlib==3.10.7
matplotlib-inline==0.2.1